+++ lib/src/bones_api_entity_db_object_gcs.dart
@@ -1,5 +1,6 @@
 import 'dart:convert' as dart_convert;
 import 'dart:io';
+import 'dart:math' as math;
 import 'dart:typed_data';
 
 import 'package:async_extension/async_extension.dart';
@@ -86,6 +87,10 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
 
   final Directory? cacheDirectory;
   final int? cacheLimit;
+  final int? cacheFilesLimit;
+
+  final int? cacheCheckMaxSkips;
+  final Duration? cacheCheckTimeout;
 
   late final gcs.Storage storage;
   late final gcs.Bucket bucket;
@@ -96,6 +101,9 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
     this.bucketName, {
     this.cacheDirectory,
     this.cacheLimit,
+    this.cacheFilesLimit,
+    this.cacheCheckMaxSkips,
+    this.cacheCheckTimeout,
     super.generateTables,
     super.populateTables,
     super.populateSource,
@@ -154,6 +162,7 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
 
     Directory? cacheDir;
     int? cacheLimit;
+    int? cacheFilesLimit;
 
     if (cacheConfig != null) {
       var path = cacheConfig.getAsString('path', ignoreCase: true)?.trim();
@@ -194,6 +203,19 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
           }
         }
       }
+
+      var filesLimitStr =
+          (cacheConfig.getAsString('filesLimit', ignoreCase: true) ??
+                  cacheConfig.getAsString('files-limit', ignoreCase: true))
+              ?.trim();
+
+      if (filesLimitStr != null && filesLimitStr.isNotEmpty) {
+        filesLimitStr = filesLimitStr.trim();
+        cacheFilesLimit = int.tryParse(filesLimitStr);
+        if (cacheFilesLimit != null && cacheFilesLimit <= 0) {
+          cacheFilesLimit = null;
+        }
+      }
     }
 
     var populate = config?['populate'];
@@ -223,6 +245,7 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
       bucket,
       cacheDirectory: cacheDir,
       cacheLimit: cacheLimit,
+      cacheFilesLimit: cacheFilesLimit,
       parentRepositoryProvider: parentRepositoryProvider,
       generateTables: generateTables,
       populateTables: populateTables,
@@ -750,7 +773,7 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
     var jsonBytes = _jsonEncoder.convert(obj);
 
     var cacheFile = _resolveCacheObjectFile(table, id);
-    if (cacheFile != null && cacheFile.existsSync()) {
+    if (cacheFile != null && await cacheFile.exists()) {
       var info = await _getObjectInfo(file);
 
       if (info != null && info.length == jsonBytes.length) {
@@ -758,14 +781,11 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
         var crc32Json = Crc32C().convert(jsonBytes).crc32CChecksumBytes;
 
         var crc32JsonOk = crc32GCS.equalsElements(crc32Json);
-
         if (crc32JsonOk) {
-          var cacheBytes = cacheFile.readAsBytesSync();
-
+          var cacheBytes = await cacheFile.readAsBytes();
           if (!cacheBytes.equalsElements(jsonBytes)) {
             cacheFile.writeAsBytesSync(jsonBytes);
           }
-
           return true;
         }
       }
@@ -776,15 +796,23 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
       jsonBytes,
       contentType: _objectContentType,
     );
+
     var ok = objInfo.length == jsonBytes.length;
 
-    cacheFile = _resolveCacheObjectFile(table, id, autoCreateDir: true);
+    cacheFile = _resolveCacheObjectFile(
+      table,
+      id,
+      autoCreateDir: true, // Ensure that the file's parent directory is created
+    );
 
     if (cacheFile != null) {
       _checkCacheDirectoryLimit();
-      _checkCacheFileParent(cacheFile);
+
+      assert(_checkCacheFileParent(cacheFile));
       cacheFile.writeAsBytesSync(jsonBytes);
+
       _checkCacheDirectoryLimitUntrackedTotal += jsonBytes.length;
+      _checkCacheDirectoryFilesLimitUntrackedTotal++;
     }
 
     return ok;
@@ -797,11 +825,10 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
 
     List<int> bytes;
 
-    if (cacheFile != null && cacheFile.existsSync()) {
-      bytes = cacheFile.readAsBytesSync();
+    if (cacheFile != null && (await cacheFile.exists())) {
+      bytes = await cacheFile.readAsBytes();
     } else {
       var file = _resolveObjectFilePath(table, id);
-
       try {
         bytes = await bucket
             .read(file)
@@ -817,6 +844,7 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
     }
 
     _checkCacheDirectoryLimitUntrackedTotal += bytes.length;
+    _checkCacheDirectoryFilesLimitUntrackedTotal++;
 
     var json = _utf8Decoder.convert(bytes);
 
@@ -824,11 +852,12 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
     return obj;
   }
 
-  void _checkCacheFileParent(File cacheFile) {
+  bool _checkCacheFileParent(File cacheFile) {
     var parent = cacheFile.parent;
     if (!parent.existsSync()) {
       parent.createSync(recursive: true);
     }
+    return true;
   }
 
   Future<Map<String, dynamic>?> _deleteObject(String table, Object? id) async {
@@ -846,38 +875,59 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
 
   DateTime _checkCacheDirectoryLimitLastTime = DateTime.utc(2020);
   int _checkCacheDirectoryLimitSkips = 0;
-  int _checkCacheDirectoryLimitLastTotal = 0;
+  int _checkCacheDirectoryLimitLastTotalSize = 0;
+  int _checkCacheDirectoryLimitLastTotalFiles = 0;
   int _checkCacheDirectoryLimitUntrackedTotal = 0;
+  int _checkCacheDirectoryFilesLimitUntrackedTotal = 0;
 
-  void _checkCacheDirectoryLimit({bool force = false}) async {
+  Future<bool> _checkCacheDirectoryLimit({bool force = false}) async {
     var cacheLimit = this.cacheLimit ?? 0;
-    if (cacheLimit <= 0) return;
+    var cacheFilesLimit = this.cacheFilesLimit ?? 0;
+    if (cacheLimit <= 0 && cacheFilesLimit <= 0) return false;
+
+    var cacheCheckMaxSkips = math.max(10, this.cacheCheckMaxSkips ?? 200);
+
+    var cacheCheckTimeout = this.cacheCheckTimeout ?? Duration(minutes: 10);
+    if (cacheCheckTimeout.inSeconds < 10) {
+      cacheCheckTimeout = Duration(seconds: 10);
+    }
 
     if (!force) {
-      var estimatedTotal =
-          _checkCacheDirectoryLimitLastTotal +
+      var estimatedTotalSize =
+          _checkCacheDirectoryLimitLastTotalSize +
           _checkCacheDirectoryLimitUntrackedTotal;
 
-      var inLimit = estimatedTotal <= cacheLimit;
-      var lowSkips = _checkCacheDirectoryLimitSkips < 100;
+      var estimatedTotalFiles =
+          _checkCacheDirectoryLimitLastTotalFiles +
+          _checkCacheDirectoryFilesLimitUntrackedTotal;
+
+      var inLimitSize = cacheLimit == 0 || estimatedTotalSize <= cacheLimit;
+      var inLimitFiles =
+          cacheFilesLimit == 0 || estimatedTotalFiles <= cacheFilesLimit;
+
+      var lowSkips = _checkCacheDirectoryLimitSkips < cacheCheckMaxSkips;
+
       var notExpired =
-          _checkCacheDirectoryLimitLastTime.elapsedTime < Duration(minutes: 5);
+          _checkCacheDirectoryLimitLastTime.elapsedTime < cacheCheckTimeout;
 
       // Skip check:
-      if (inLimit && lowSkips && notExpired) {
+      if (inLimitSize && inLimitFiles && lowSkips && notExpired) {
         ++_checkCacheDirectoryLimitSkips;
-        return;
+        return false;
       }
     }
 
     var list = cacheDirectory?.listSync(recursive: true);
-    if (list == null || list.isEmpty) return;
+    if (list == null || list.isEmpty) return false;
 
-    _checkCacheDirectoryLimitLastTime = DateTime.now();
+    final checkInitTime = DateTime.now();
+    _checkCacheDirectoryLimitLastTime = checkInitTime;
     _checkCacheDirectoryLimitSkips = 0;
 
     var files =
         list.whereType<File>().where((f) => f.path.endsWith(".json")).toList();
+
+    var totalFiles = files.length;
 
     var filesStats =
         await files
@@ -887,24 +937,38 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
 
     var totalSize = filesStats.values.map((s) => s.size).sum;
 
-    _checkCacheDirectoryLimitLastTotal = totalSize;
+    _checkCacheDirectoryLimitLastTotalSize = totalSize;
+    _checkCacheDirectoryLimitLastTotalFiles = totalFiles;
     _checkCacheDirectoryLimitUntrackedTotal = 0;
+    _checkCacheDirectoryFilesLimitUntrackedTotal = 0;
 
-    if (totalSize <= cacheLimit) {
+    final checkTime = DateTime.now().difference(checkInitTime);
+
+    final cacheLimitOk = cacheLimit == 0 || totalSize <= cacheLimit;
+    final cacheFilesLimitOk =
+        cacheFilesLimit == 0 || totalFiles <= cacheFilesLimit;
+
+    if (cacheLimitOk && cacheFilesLimitOk) {
       var r = (totalSize / cacheLimit) * 100;
       _log.info(
-        '[CACHE] Limit check: OK ($totalSize / $cacheLimit bytes ${r.toStringAsFixed(2)}%)  (${filesStats.length} files)',
+        '[CACHE] Limit check[${checkTime.inMilliseconds} ms]: OK ($totalSize${cacheLimit > 0 ? ' / $cacheLimit' : ''} bytes ${r.toStringAsFixed(2)}%)  ($totalFiles${cacheFilesLimit > 0 ? ' / $cacheFilesLimit' : ''} files)',
       );
-      return;
+      return false;
     }
 
     var entries = filesStats.entries.toList();
     entries.sort((a, b) => a.value.modified.compareTo(b.value.modified));
 
-    final delNeededSize = ((totalSize * 0.80) - cacheLimit).toInt();
+    final delNeededFiles =
+        cacheFilesLimit <= 0
+            ? 0
+            : ((totalFiles * 0.80) - cacheFilesLimit).toInt();
+
+    final delNeededSize =
+        cacheLimit <= 0 ? 0 : ((totalSize * 0.80) - cacheLimit).toInt();
 
     _log.info(
-      '[CACHE] Reached limit: $totalSize / $cacheLimit bytes (${entries.length} files)! Releasing $delNeededSize bytes...',
+      '[CACHE] Limit Check[${checkTime.inMilliseconds} ms] reached limit: $totalSize${cacheLimit > 0 ? ' / $cacheLimit' : ''} bytes (${entries.length}${cacheFilesLimit > 0 ? ' / $cacheFilesLimit' : ''} files)! Releasing $delNeededSize bytes...',
     );
 
     var del = <File>[];
@@ -915,16 +979,24 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
       var s = e.value;
 
       del.add(f);
-
       delSize += s.size;
-      if (delSize >= delNeededSize) break;
+
+      if (delSize >= delNeededSize && del.length >= delNeededFiles) break;
     }
 
-    _log.info('[CACHE] Removing ${del.length} files...');
+    _log.info('[CACHE] Removing ${del.length} files... ($delSize bytes)');
 
-    var delResuls = await del.map((f) => f.delete()).toList().resolveAll();
+    final removeInitTime = DateTime.now();
 
-    _log.info('[CACHE] Removed files: ${delResuls.length} ($delSize bytes)');
+    var delResults = await del.map((f) => f.delete()).toList().resolveAll();
+
+    final removeTime = DateTime.now().difference(removeInitTime);
+
+    _log.info(
+      '[CACHE] Removed files[${removeTime.inMilliseconds} ms]: ${delResults.length} ($delSize bytes)',
+    );
+
+    return true;
   }
 
   bool _isValidId(Object? id) {
@@ -955,6 +1027,7 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
       autoCreateDir: autoCreateDir,
     );
     if (tableDir == null) return null;
+
     var idStr = _normalizeID(id);
     var file = File(pack_path.join(tableDir.path, '$idStr.json'));
     return file;
@@ -966,6 +1039,7 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
   }) {
     var cacheDirectory = this.cacheDirectory;
     if (cacheDirectory == null) return null;
+
     var tableStr = _normalizeTableName(table);
     var tableDir = Directory(pack_path.join(cacheDirectory.path, tableStr));